Ok so I forgot that we need to use numeric object arrays for this part too

+++ src/frontend/edu/brown/utils/ParameterMangler.java
@@ -106,7 +106,7 @@ public class ParameterMangler {
                 try {
                     switch (this.param_types[i]) {
                         case TINYINT: {
-                            byte arr[] = (byte[]) orig[i];
+                            Byte arr[] = (Byte[]) orig[i];
                             inner = new Object[arr.length];
                             for (int j = 0; j < arr.length; j++) {
                                 inner[j] = arr[j];
@@ -114,7 +114,7 @@ public class ParameterMangler {
                             break;
                         }
                         case SMALLINT: {
-                            short arr[] = (short[]) orig[i];
+                            Short arr[] = (Short[]) orig[i];
                             inner = new Object[arr.length];
                             for (int j = 0; j < arr.length; j++) {
                                 inner[j] = arr[j];
@@ -122,7 +122,7 @@ public class ParameterMangler {
                             break;
                         }
                         case INTEGER: {
-                            int arr[] = (int[]) orig[i];
+                            Integer arr[] = (Integer[]) orig[i];
                             inner = new Object[arr.length];
                             for (int j = 0; j < arr.length; j++) {
                                 inner[j] = arr[j];
@@ -130,7 +130,7 @@ public class ParameterMangler {
                             break;
                         }
                         case BIGINT: {
-                            long arr[] = (long[]) orig[i];
+                            Long arr[] = (Long[]) orig[i];
                             inner = new Object[arr.length];
                             for (int j = 0; j < arr.length; j++) {
                                 inner[j] = arr[j];
@@ -138,7 +138,7 @@ public class ParameterMangler {
                             break;
                         }
                         case FLOAT: {
-                            float arr[] = (float[]) orig[i];
+                            Double arr[] = (Double[]) orig[i];
                             inner = new Object[arr.length];
                             for (int j = 0; j < arr.length; j++) {
                                 inner[j] = arr[j];
